Refactor(structure): Component exports and imports
