+++ src/components/Form/Settings/Checkbox/Checkbox.jsx
@@ -58,4 +58,4 @@ class Checkbox extends PureComponent {
   }
 }
 
-export default Checkbox;
+export { Checkbox as default, Checkbox };
