+++ src/components/Form/Settings/Checkbox/index.jsx
@@ -1,0 +1,1 @@
+export * from './Checkbox';
