+++ src/components/Form/Settings/ChipSelect/index.jsx
@@ -1,0 +1,1 @@
+export * from './ChipSelect';
