+++ src/components/Form/Settings/FileUpload/index.jsx
@@ -1,0 +1,1 @@
+export * from './FileUpload';
