+++ src/components/Form/Settings/Radio/index.jsx
@@ -1,0 +1,1 @@
+export * from './Radio';
