+++ src/components/Form/Settings/Slider/index.jsx
@@ -1,0 +1,1 @@
+export * from './Slider';
