+++ src/components/Form/Settings/Switch/index.jsx
@@ -1,0 +1,1 @@
+export * from './Switch';
