+++ src/components/Form/Settings/Text/index.jsx
@@ -1,0 +1,1 @@
+export * from './Text';
