+++ src/components/Form/Settings/index.jsx
@@ -1,0 +1,8 @@
+export * from './Checkbox';
+export * from './ChipSelect';
+export * from './Dropdown';
+export * from './FileUpload';
+export * from './Radio';
+export * from './Slider';
+export * from './Switch';
+export * from './Text';
