+++ src/components/Form/index.jsx
@@ -1,0 +1,1 @@
+export * from './Settings';
