+++ src/components/Layout/Settings/Header/index.jsx
@@ -1,0 +1,1 @@
+export * from './Header';
