+++ src/components/Layout/Settings/Item/index.jsx
@@ -1,0 +1,1 @@
+export * from './SettingsItem';
