+++ src/components/Layout/Settings/PreferencesWrapper/PreferencesWrapper.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import { Row, Content, Action } from '../../../../components/Layout/Settings/Item/SettingsItem';
+import { Row, Content, Action } from '../Item/SettingsItem';
 import variables from 'config/variables';
-import Slider from '../../../../components/Form/Settings/Slider/Slider';
+import Slider from '../../../Form/Settings/Slider/Slider';
 
 import { values } from 'modules/helpers/settings/modals';
 import EventBus from 'modules/helpers/eventbus';
