+++ src/components/Layout/Settings/PreferencesWrapper/index.jsx
@@ -1,0 +1,1 @@
+export * from './PreferencesWrapper';
