+++ src/components/Layout/Settings/index.jsx
@@ -1,0 +1,4 @@
+export * from './Header';
+export * from './Item';
+export * from './PreferencesWrapper';
+export * from './Section';
