+++ src/components/Layout/index.jsx
@@ -1,1 +1,2 @@
+export * from './Settings';
 export * from './WidgetsLayout';
