+++ src/features/modals/main/marketplace/sections/Marketplace.jsx
@@ -13,7 +13,7 @@ import {
 import Item from '../Item';
 import Items from '../Items';
 import Dropdown from '../../../../../components/Form/Settings/Dropdown/Dropdown';
-import Header from '../../../../../components/Layout/Settings/Header/Header';
+import { Header } from 'components/Layout/Settings';
 import { Button } from 'components/Elements';
 
 import { install, urlParser, uninstall } from 'modules/helpers/marketplace';
