+++ src/features/modals/main/settings/sections/Greeting.jsx
@@ -1,13 +1,13 @@
 import variables from 'config/variables';
 import { useState } from 'react';
 
-import Header from '../../../../../components/Layout/Settings/Header/Header';
+import { Header } from 'components/Layout/Settings';
 import Checkbox from '../../../../../components/Form/Settings/Checkbox/Checkbox';
 import Switch from '../../../../../components/Form/Settings/Switch/Switch';
 import Text from '../../../../../components/Form/Settings/Text/Text';
 
 import { Row, Content, Action } from '../../../../../components/Layout/Settings/Item/SettingsItem';
-import PreferencesWrapper from '../PreferencesWrapper';
+import PreferencesWrapper from '../../../../../components/Layout/Settings/PreferencesWrapper/PreferencesWrapper';
 
 const GreetingSettings = () => {
   const [birthday, setBirthday] = useState(
