+++ src/features/modals/main/settings/sections/Search.jsx
@@ -3,7 +3,7 @@ import { PureComponent } from 'react';
 import { toast } from 'react-toastify';
 import { MenuItem, TextField } from '@mui/material';
 
-import Header from '../../../../../components/Layout/Settings/Header/Header';
+import { Header } from 'components/Layout/Settings';
 import Dropdown from '../../../../../components/Form/Settings/Dropdown/Dropdown';
 import Checkbox from '../../../../../components/Form/Settings/Checkbox/Checkbox';
 
@@ -12,7 +12,7 @@ import { Row, Content, Action } from '../../../../../components/Layout/Settings/
 import EventBus from 'modules/helpers/eventbus';
 
 import searchEngines from 'features/widgets/search/search_engines.json';
-import PreferencesWrapper from '../PreferencesWrapper';
+import PreferencesWrapper from '../../../../../components/Layout/Settings/PreferencesWrapper/PreferencesWrapper';
 
 export default class SearchSettings extends PureComponent {
   constructor() {
