+++ src/features/modals/main/settings/sections/Weather.jsx
@@ -3,13 +3,13 @@ import { PureComponent } from 'react';
 
 import { MdAutoAwesome } from 'react-icons/md';
 
-import Header from '../../../../../components/Layout/Settings/Header/Header';
+import { Header } from 'components/Layout/Settings';
 import Radio from '../../../../../components/Form/Settings/Radio/Radio';
 import Dropdown from '../../../../../components/Form/Settings/Dropdown/Dropdown';
 import Checkbox from '../../../../../components/Form/Settings/Checkbox/Checkbox';
 import { TextField } from '@mui/material';
 
-import PreferencesWrapper from '../PreferencesWrapper';
+import PreferencesWrapper from '../../../../../components/Layout/Settings/PreferencesWrapper/PreferencesWrapper';
 
 import { Row, Content, Action } from '../../../../../components/Layout/Settings/Item/SettingsItem';
